requirements.txt
dotenv==0.9.9
psycopg2-binary==2.9.10
python-dotenv==1.1.0
websockets==15.0.1
